Implementar acumulado de totales en factura y metodo de carga CSV

+++ src/main/java/com/tienda/backend/entity/ProductosEntity.java
@@ -27,7 +27,7 @@ public class ProductosEntity {
     private String nombreProducto;
 
     @Column(name = "precio")
-    private int precio;
+    private double precio;
 
     @OneToMany(mappedBy = "productosEnt", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     @JsonIgnore
